+++ testdata/apps/kindtree/app/formularios/fundo/route.go
@@ -1,0 +1,5 @@
+package fundo
+
+import "github.com/emersonjoe/trilha"
+
+func POST(c *trilha.Ctx) error { return c.Text(200, "ok") }