+++ testdata/apps/kindtree/app/paginas/page.go
@@ -1,0 +1,8 @@
+package paginas
+
+import (
+	"github.com/emersonjoe/trilha"
+	"github.com/emersonjoe/trilha/h"
+)
+
+func Page(c *trilha.Ctx) (h.Node, error) { return h.Div(), nil }